refactor/874 - Combine Pilot tests
