--- test/aircraft/Pilot/_baseSpec.tmp
@@ -1,18 +1,0 @@
-/* eslint-disable max-len */
-// Use this file as a starter for future Pilot tests that get split out by the method
-// TODO: remove this file once a majority of the tests have been added.
-
-import ava from 'ava';
-import sinon from 'sinon';
-import _isArray from 'lodash/isArray';
-import _isEqual from 'lodash/isEqual';
-import _isObject from 'lodash/isObject';
-
-import Pilot from '../../../src/assets/scripts/client/aircraft/Pilot/Pilot';
-import {
-    fmsArrivalFixture,
-    modeControllerFixture
-} from '../../fixtures/aircraftFixtures';
-
-const runwayMock = '19L';
-const cruiseSpeedMock = 460;
